Watch product name value change in inventory section
